requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

Flask
paramiko
azure-functions
azure-identity
azure-storage-blob
google-cloud-storage
